Revert "big refactor"
This reverts commit 23dc8330
