+++ iOSBot.Bot/IosBot.cs
@@ -1,15 +1,11 @@
-﻿using System.Collections.Concurrent;
-using System.Timers;
-using Discord;
+﻿using Discord;
 using Discord.Rest;
 using Discord.WebSocket;
-using iOSBot.Bot.Commands;
-using iOSBot.Data;
+using iOSBot.Bot.Singletons;
 using iOSBot.Service;
 using Microsoft.Extensions.DependencyInjection;
 using Newtonsoft.Json.Linq;
 using NLog;
-using Timer = System.Timers.Timer;
 
 namespace iOSBot.Bot
 {
@@ -21,44 +17,29 @@ namespace iOSBot.Bot
 
         private readonly IServiceProvider _serviceProvider;
 
-        private DiscordSocketClient Client { get; set; }
-        private DiscordRestClient RestClient { get; set; }
+        private DiscordSocketClient? Client { get; set; }
+        private DiscordRestClient? RestClient { get; set; }
 
-        private IDiscordService Discord { get; set; }
-        private IAppleService Apple { get; set; }
-        
-        public Timer Timer { get; set; }
+        private readonly ApiSingleton _apiFeed = ApiSingleton.Instance;
 
         private string? Status { get; set; }
 
         public IosBot()
         {
-            _serviceProvider = CreateProvider() ?? throw new Exception("cannot create service provider");
-
-            Discord = _serviceProvider.GetService<IDiscordService>() ?? throw new Exception("Cannot create Discord service");
-            Apple = _serviceProvider.GetService<IAppleService>() ?? throw new Exception("Cannot create Apple service");;
-            
-            Timer = new Timer
-            {
-                Interval = 90 * 1000 // 90 seconds
-            };
-            
-            Timer.Elapsed += TimerOnElapsed;
-            Timer.Start();
+            _serviceProvider = CreateProvider();
         }
 
         private IServiceProvider CreateProvider()
         {
             var config = new DiscordSocketConfig()
             {
-                GatewayIntents = GatewayIntents.GuildMessages | GatewayIntents.DirectMessages | GatewayIntents.GuildMembers,
+                GatewayIntents = GatewayIntents.GuildMessages | GatewayIntents.DirectMessages,
                 MessageCacheSize = 15
             };
 
             var collection = new ServiceCollection();
 
-            collection.AddScoped<IAppleService, AppleService>();
-            collection.AddScoped<IDiscordService, DiscordService>();
+            collection.AddTransient<IAppleService, AppleService>();
             collection
                 .AddSingleton(config)
                 .AddSingleton<DiscordSocketClient>();
@@ -71,7 +52,7 @@ namespace iOSBot.Bot
         private async Task MainAsync(string[] args)
         {
             AppDomain.CurrentDomain.UnhandledException += CurrentDomain_UnhandledException;
-            
+
             Client = _serviceProvider.GetRequiredService<DiscordSocketClient>();
             RestClient = Client.Rest;
 
@@ -87,7 +68,7 @@ namespace iOSBot.Bot
                     Status = "in testing mode";
                     break;
             }
-            
+
             await Client.LoginAsync(TokenType.Bot, args[0]);
             await Client.SetGameAsync(Status, type: ActivityType.Watching);
 
@@ -99,23 +80,24 @@ namespace iOSBot.Bot
             Client.MessageReceived += _client_MessageReceived;
             Client.JoinedGuild += ClientOnJoinedGuild;
 
+            _apiFeed.Bot = Client;
+            _apiFeed.Start();
+
             await Task.Delay(-1);
         }
 
         private Task ClientOnJoinedGuild(SocketGuild arg)
         {
-            using var db = new BetaContext();
-
-            Discord.PostToServers(Client, db.ErrorServers.Select(s => s.ServerId), $"Joined Server: {arg.Name}");
+            Commands.PostError(Client, new AppleService(), $"Joined Server {arg.Name}.");
             return Task.CompletedTask;
         }
 
         private Task _client_MessageReceived(SocketMessage arg)
         {
-            if (arg.Channel is not IDMChannel || arg.Author.Id == Client.GetApplicationInfoAsync().Result.Id) return Task.CompletedTask;
-            using var db = new BetaContext();
-            
-            Discord.PostToServers(Client, db.ErrorServers.Select(s => s.ServerId), $"DM Received:\n{arg.Content}\n-@{arg.Author}");
+            if (arg.Channel is not IDMChannel || arg.Author.Id == Client.GetApplicationInfoAsync().Result.Id)
+                return Task.CompletedTask;
+
+            Commands.PostError(Client, new AppleService(), $"DM Received:\n{arg.Content}\n-@{arg.Author}");
             arg.Channel.SendMessageAsync("Sending this message along. thank you.");
 
             return Task.CompletedTask;
@@ -123,10 +105,8 @@ namespace iOSBot.Bot
 
         private void CurrentDomain_UnhandledException(object sender, UnhandledExceptionEventArgs e)
         {
-            using var db = new BetaContext();
-
             _logger.Error($"{e}\n{e.ExceptionObject}");
-            Discord.PostToServers(Client, db.ErrorServers.Select(s => s.ServerId), $"Unhandled Exception:\n{e}\n{e.ExceptionObject}");
+            Commands.PostError(Client, new AppleService(), $"Unhandled Exception:\n{e}\n{e.ExceptionObject}");
         }
 
         private Task _client_SlashCommandExecuted(SocketSlashCommand arg)
@@ -136,81 +116,65 @@ namespace iOSBot.Bot
             {
                 jsonArgs.Add(new JProperty(o.Name, o.Value.ToString()));
             }
-            _logger.Info($"Command received: {arg.CommandName} in {RestClient.GetChannelAsync(arg.ChannelId!.Value).Result} from {arg.User.Username}\n```json\nargs:{jsonArgs}\n```");
+
+            _logger.Info(
+                $"Command received: {arg.CommandName} in {RestClient.GetChannelAsync(arg.ChannelId!.Value).Result} from {arg.User.Username}\n```json\nargs:{jsonArgs}\n```");
 
             switch (arg.CommandName)
             {
                 case "watch":
-                    AppleCommands.YesWatch(arg, RestClient);
+                    Commands.InitCommand(arg, RestClient);
                     break;
                 case "unwatch":
-                    AppleCommands.NoWatch(arg, RestClient);
+                    Commands.RemoveCommand(arg, RestClient);
                     break;
                 case "force":
-                    if (!AdminCommands.IsAllowed(arg.User.Id))
-                    {
-                        arg.RespondAsync("Only the bot creator can use this command.", ephemeral: true);
-                    }
-
-                    Timer.Stop();
-                    TimerOnElapsed(null, null!);
-                    Timer.Start();
-                    arg.RespondAsync("Updates Checked", ephemeral: true);
+                    Commands.ForceCommand(arg);
                     break;
                 case "error":
-                    AdminCommands.YesErrors(arg, RestClient);
+                    Commands.ErrorCommand(arg, RestClient);
                     break;
                 case "noerror":
-                    AdminCommands.NoErrors(arg);
+                    Commands.RemoveErrorCommand(arg);
                     break;
                 case "update":
-                    AdminCommands.UpdateOptions(arg, Client);
+                    Commands.UpdateOptions(arg, Client);
                     break;
                 case "manifest":
-                    MemeCommands.Manifest(arg);
+                    Commands.Manifest(arg);
                     break;
                 case "goodbot":
-                    MemeCommands.GoodBot(arg, RestClient);
+                    Commands.GoodBot(arg, RestClient);
                     break;
                 case "badbot":
-                    MemeCommands.BadBot(arg, RestClient);
+                    Commands.BadBot(arg, RestClient);
                     break;
                 case "info":
-                    AppleCommands.DeviceInfo(arg);
+                    Commands.DeviceInfo(arg);
+                    break;
+                case "when":
+                    Commands.When(arg);
                     break;
                 case "servers":
-                    AdminCommands.GetServers(arg, Client);
+                    Commands.GetServers(arg, RestClient);
                     break;
                 case "status":
-                    string response = Timer.Enabled ? "Running" : "Not Running";
-                    arg.RespondAsync(ephemeral: true, text: $"Bot is currently: {response}");
+                    Commands.BotStatus(arg, StatusCommand.STATUS);
                     break;
                 case "start":
-                    if (!AdminCommands.IsAllowed(arg.User.Id))
-                    {
-                        arg.RespondAsync("Only the bot creator can use this command.", ephemeral: true);
-                    }
-
-                    Timer.Start();
-                    arg.FollowupAsync(ephemeral: true, text: $"Bot is running");
+                    Commands.BotStatus(arg, StatusCommand.START);
                     break;
                 case "stop":
-                    if (!AdminCommands.IsAllowed(arg.User.Id))
-                    {
-                        arg.RespondAsync("Only the bot creator can use this command.", ephemeral: true);
-                    }
-
-                    Timer.Stop();
-                    arg.FollowupAsync(ephemeral: true, text: $"Bot is stopped");
+                    Commands.BotStatus(arg, StatusCommand.STOP);
                     break;
                 case "yesthreads":
-                    AppleCommands.NewThreadChannel(arg);
+                    Commands.NewThreadChannel(arg);
                     break;
                 case "nothreads":
-                    AppleCommands.DeleteThreadChannel(arg);
+                    Commands.DeleteThreadChannel(arg);
                     break;
                 case "fake":
-                    AdminCommands.FakeUpdate(arg, Discord);
+                    Commands.FakeUpdate(arg);
                     break;
                 default:
                     break;
@@ -221,7 +185,7 @@ namespace iOSBot.Bot
 
         private async Task _client_Ready()
         {
-            CommandInitializer.UpdateCommands(Client);
+            Commands.UpdateCommands(Client);
         }
 
         private async Task _client_Log(LogMessage arg)
@@ -229,83 +193,10 @@ namespace iOSBot.Bot
             _logger.Info(arg.Message);
             if (null != arg.Exception)
             {
-                // dont care about logging these
-                if (arg.Exception.Message.EndsWith("Server requested a reconnect") ||
-                    arg.Exception.Message.EndsWith("WebSocket connection was closed")) return;
-                
-                using var db = new BetaContext();
-                Discord.PostToServers(Client, db.ErrorServers.Select(s => s.ChannelId), $"Bot error:\n{arg.Exception.Message}");
+                Commands.PostError(Client, new AppleService(), $"Bot error:\n{arg.Exception.Message}");
                 _logger.Error(arg.Exception);
                 _logger.Error(arg.Exception.InnerException.StackTrace);
             }
-        }
-        
-        private async void TimerOnElapsed(object? sender, ElapsedEventArgs e)
-        {
-            var updates = new ConcurrentBag<Service.Update>();
-            using var db = new BetaContext();
-            var dbUpdates = db.Updates.ToList();
-
-            // check all devices for /info 
-            Parallel.ForEach(db.Devices, device =>
-            {
-                try
-                {
-                    Service.Update u = Apple.GetUpdate(device).Result;
-                    
-                    if (!dbUpdates.Any(up => up.Build == u.Build &&
-                                             up.ReleaseDate == u.ReleaseDate &&
-                                             up.Category == u.Group)) updates.Add(u);
-                    else _logger.Info("No new update found for " + device.FriendlyName);
-                }
-                catch (Exception ex)
-                {
-                    Discord.PostToServers(Client, db.ErrorServers.Select(s => s.ChannelId), $"Error checking update for {device.FriendlyName}:\n{ex.Message}");
-                }
-            });
-            
-            foreach (var update in updates)
-            {
-                var category = update.Device.Category;
-                var servers = db.Servers.Where(s => s.Category == category);
-                var threads = db.Threads.Where(t => t.Category == category);
-
-                _logger.Info($"Update for {update.Device.FriendlyName} found. Version {update.VersionReadable} with build id {update.Build}");
-
-                // dont post if older than 12 hours, still add to db tho
-                var postOld = Convert.ToBoolean(Convert.ToInt16(db.Configs.First(c => c.Name == "PostOld").Value));
-                
-                if (postOld || update.ReleaseDate.DayOfYear == DateTime.Today.DayOfYear)
-                {    
-                    foreach (var server in servers)
-                    {
-                        Discord.PostUpdate(update, server);
-                    }
-
-                    foreach (var thread in threads)
-                    {
-                        Discord.PostThread(update, thread);
-                    }
-                }
-                else
-                {
-                    var error = $"{update.Device.FriendlyName} update {update.VersionReadable}-{update.Build} was released on {update.ReleaseDate.ToShortDateString()}. too old. not posting.";
-                    _logger.Info(error);
-                    Discord.PostToServers(Client, db.ErrorServers.Select(s => s.ChannelId), error);
-                }
-                
-                Apple.SaveUpdate(update);
-            }
-            
-            // update server count
-            ulong channelId = ulong.Parse(db.Configs.First(c => c.Name == "StatusChannel").Value);
-            string env = db.Configs.First(c => c.Name == "Environment").Value;
-            IChannel channel = Client.GetChannelAsync(channelId).Result;
-            await ((IVoiceChannel)channel).ModifyAsync(c => c.Name = $"{env} Bot Servers: {RestClient.GetGuildsAsync().Result.Count}");
-
-            Timer.Interval = int.Parse(db.Configs.First(c => c.Name == "Timer").Value);
-
-            await db.SaveChangesAsync();
         }
     }
 }
